[BUILD] Add StartupSarosSTF to start the STF
I added a StartupSarosSTF.java and a plugin.xml to the stf project. Now
the STF starts from the stf-project and not from the eclipse-project
like it was befor.
