+++ de.fu_berlin.inf.dpp.stf.test/META-INF/MANIFEST.MF
