+++ de.fu_berlin.inf.dpp.stf/build.properties
@@ -4,10 +4,15 @@ source.. = src/html/,\
 output.. = bin/
 bin.includes = META-INF/,\
                .,\
+               plugin.xml,\
                lib/swt-browser-improved.jar,\
                lib/org.eclipse.swtbot.eclipse.finder_2.0.5.20110922_1046-258a521-dev-e36.jar,\
                lib/org.eclipse.swtbot.swt.finder_2.0.5.20110922_1046-258a521-dev-e36.jar,\
                lib/org.eclipse.swtbot.eclipse.gef.finder_2.0.5.20110922_1046-258a521-dev-e36.jar,\
-               lib/org.hamcrest.library_1.1.0.v20090501071000.jar
+               lib/org.hamcrest_1.1.0.v20090501071000.jar,\
+               lib/org.hamcrest.text_1.1.0.v20090501071000.jar,\
+               lib/org.hamcrest.library_1.1.0.v20090501071000.jar,\
+               lib/org.hamcrest.integration_1.1.0.v20090501071000.jar,\
+               lib/org.hamcrest.core_1.1.0.v20090501071000.jar
 src.excludes = test/html/,\
                test/eclipse/
